GH-87 Extract record test fixture for readability
